Changes and additions 2/21

+++ Lab1/src/pkgCore/Die.java
@@ -2,14 +2,14 @@ package pkgCore;
 
 public class Die {
 
-	private byte DieValue;
+	private int DieValue;
 
 	public Die() {
 		// TODO: Determine DieVaue.. a random number between 1 and 6
-		DieValue = (byte)(Math.random()*6 + 1);			
+		DieValue = (int)(Math.random()*6 + 1);			
 	}
 
-	public byte getDieValue() {
+	public int getDieValue() {
 		return DieValue;
 	}
 }
